requirements.txt
autopep8==2.0.4
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
feedgen==0.9.0
Flask==2.3.3
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
pycodestyle==2.11.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
validators==0.22.0
Werkzeug==2.3.7
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
